# sstable: configure bufio.Writer with 512 KB buffer