Previously, sstable.NewWriter used the default bufio.NewWriter which
uses a 4 KB buffer. This change updates sstable.NewWriter to construct a
bufio.Writer with a 512 KB buffer.

```
name                           old time/op    new time/op    delta
WriterSyncFile/vsize=64-16        145ns ± 8%     119ns ± 7%   -17.79%  (p=0.000 n=10+10)
WriterSyncFile/vsize=512-16       350ns ± 7%     212ns ± 4%   -39.51%  (p=0.000 n=10+10)
WriterSyncFile/vsize=1024-16      572ns ± 6%     320ns ± 6%   -43.95%  (p=0.000 n=10+9)
WriterSyncFile/vsize=2048-16     1.02µs ± 7%    0.52µs ± 3%   -48.68%  (p=0.000 n=10+9)
WriterSyncFile/vsize=4096-16     1.89µs ± 6%    0.92µs ± 6%   -51.28%  (p=0.000 n=10+9)
WriterSyncFile/vsize=8192-16     3.62µs ± 7%    1.73µs ± 4%   -52.19%  (p=0.000 n=10+9)
WriterSyncFile/vsize=16384-16    7.20µs ± 7%    3.38µs ± 4%   -53.12%  (p=0.000 n=10+9)
WriterSyncFile/vsize=32768-16    14.3µs ±13%     6.6µs ± 5%   -53.62%  (p=0.000 n=10+9)
WriterSyncFile/vsize=65536-16    27.2µs ± 7%    11.8µs ± 3%   -56.49%  (p=0.000 n=10+9)

name                           old speed      new speed      delta
WriterSyncFile/vsize=64-16      443MB/s ± 8%   538MB/s ± 6%   +21.41%  (p=0.000 n=10+10)
WriterSyncFile/vsize=512-16    1.47GB/s ± 6%  2.42GB/s ± 4%   +65.05%  (p=0.000 n=10+10)
WriterSyncFile/vsize=1024-16   1.79GB/s ± 6%  3.20GB/s ± 5%   +78.16%  (p=0.000 n=10+9)
WriterSyncFile/vsize=2048-16   2.02GB/s ± 7%  3.93GB/s ± 3%   +94.59%  (p=0.000 n=10+9)
WriterSyncFile/vsize=4096-16   2.17GB/s ± 6%  4.45GB/s ± 6%  +105.10%  (p=0.000 n=10+9)
WriterSyncFile/vsize=8192-16   2.27GB/s ± 6%  4.73GB/s ± 4%  +108.89%  (p=0.000 n=10+9)
WriterSyncFile/vsize=16384-16  2.28GB/s ± 7%  4.86GB/s ± 4%  +113.08%  (p=0.000 n=10+9)
WriterSyncFile/vsize=32768-16  2.30GB/s ±12%  4.94GB/s ± 5%  +114.98%  (p=0.000 n=10+9)
WriterSyncFile/vsize=65536-16  2.41GB/s ± 7%  5.54GB/s ± 3%  +129.56%  (p=0.000 n=10+9)
```

The impact is also visible in large-value `pebble bench ycsb`
benchmarks.

```
pebble bench ycsb data-peb --duration '10m' --workload=F --values 64000 --concurrency=4
```

Old:
```
____optype__elapsed_____ops(total)___ops/sec(cum)__keys/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
insert_100   600.0s          73779          123.0          123.0     32.5     23.1     75.5     96.5    184.5
```

New:
```
____optype__elapsed_____ops(total)___ops/sec(cum)__keys/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)
insert_100   600.0s          91820          153.0          153.0     26.1     22.0     56.6     71.3    302.0
```