@jbowens 2020-06-04
Dang. On a roachprod machine using a localssd, this change was no good.